backend/requirements.txt
asgiref==3.8.1
Django==4.2.15
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
django_redis==4.12.1
Faker==28.0.0
psycopg2==2.9.4
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
six==1.16.0
sqlparse==0.5.1
stripe==10.9.0
